Fixed minor docs section detail
While reading through [the inclusion and exclusion rules config feature docs page](https://espanso.org/docs/configuration/include-and-exclude/), I observed something that could probably be improved. I went ahead and fixed it already. Feel free to ignore this if this has already been corrected in a staging branch or if it's irrelevant.

Under the "Typical use-case for include rules" heading the the passage read:
> "and only enable when when inside Chrome"

That's confusing and likely a minor and reasonable editing oversight.

> "and only enable when when inside Chrome"

Using a spelling and grammar validation feature in a modern text editor application like Google Docs, suggests the following improvement, which is the reasonable sentence the documentation editor likely intended:

> "and only enable them when inside Chrome"

Cheers!
Arty